menumind-backend/app/serializers.py

from rest_framework import serializers
from .models import Restaurant, Dish, Sales, Prediction
class SalesSerializer(serializers.ModelSerializer):
    class Meta:
        model = Sales
        fields = '__all__'
class PredictionSerializer(serializers.ModelSerializer):
    class Meta:
        model = Prediction
        fields = '__all__'
