menumind-backend/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-authtoken
redis
django-cors-headers
python-dotenv
